requirements.txt
Django==1.10.5
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.2
django-compressor==2.1
djangorestframework==3.5.3
drf-nested-routers==0.9.0
gunicorn==19.1.1
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
